Split main.rs to terminal and style files.
# Conflicts:
#	src/main.rs
