+++ src/ui/style.rs
@@ -1,0 +1,9 @@
+use ratatui::style::palette::tailwind;
+use ratatui::style::Color;
+
+pub(crate) const APP_HEADER_BG: Color = tailwind::BLACK;
+pub(crate) const NORMAL_ROW_COLOR: Color = tailwind::SLATE.c950;
+pub(crate) const ALT_ROW_COLOR: Color = tailwind::SLATE.c900;
+pub(crate) const SELECTED_STYLE_FG: Color = tailwind::YELLOW.c700;
+pub(crate) const TEXT_COLOR: Color = tailwind::SLATE.c200;
+pub(crate) const NOT_AVAILABLE_TEXT_COLOR: Color = tailwind::RED.c500;
